LayoutTests/platform/mac-monterey/imported/w3c/web-platform-tests/html/cross-origin-embedder-policy/coep-on-response-from-service-worker.https-expected.txt

CONSOLE MESSAGE: FetchEvent.respondWith received an error: TypeError: Load failed
CONSOLE MESSAGE: Fetch API cannot load https://127.0.0.1:9443/html/cross-origin-embedder-policy/resources/nothing-same-origin-corp.txt.
CONSOLE MESSAGE: Cancelled load to https://127.0.0.1:9443/common/blank.html because it violates the resource's Cross-Origin-Resource-Policy response header.
CONSOLE MESSAGE: Fetch API cannot load https://127.0.0.1:9443/common/blank.html due to access control checks.
CONSOLE MESSAGE: Cancelled load to https://127.0.0.1:9443/html/cross-origin-embedder-policy/resources/nothing-same-origin-corp.txt?passthrough because it violates the resource's Cross-Origin-Resource-Policy response header.
CONSOLE MESSAGE: Fetch API cannot load https://127.0.0.1:9443/html/cross-origin-embedder-policy/resources/nothing-same-origin-corp.txt?passthrough due to access control checks.
CONSOLE MESSAGE: Cancelled load to https://127.0.0.1:9443/common/blank.html?passthrough because it violates the resource's Cross-Origin-Resource-Policy response header.
CONSOLE MESSAGE: Fetch API cannot load https://127.0.0.1:9443/common/blank.html?passthrough due to access control checks.
CONSOLE MESSAGE: FetchEvent.respondWith received an error: TypeError: Load failed
CONSOLE MESSAGE: Fetch API cannot load https://127.0.0.1:9443/common/blank.html.

PASS setup
PASS making a same-origin request for CORP: same-origin
PASS making a same-origin request for no CORP
PASS making a same-origin request for CORP: cross-origin
PASS making a cross-origin request for CORP: same-origin
PASS making a cross-origin request for no CORP
PASS making a cross-origin request for CORP: cross-origin
PASS making a cross-origin request for CORP: same-origin [PASS THROUGH]
PASS making a cross-origin request for no CORP [PASS THROUGH]
PASS making a cross-origin request for CORP: cross-origin [PASS THROUGH]
PASS making a cross-origin request with CORS without ACAO
PASS making a cross-origin request with CORS
PASS teardown
